+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Fixed Airtable import bug where the import would fail if a row is empty.
+
 ## Released (2022-03-03 1.9.1)
 
 * Fixed bug when importing a formula or lookup field with an incorrect empty value.